fixed_bug_host_deletion

+++ apps/hosts/views.py
@@ -152,7 +152,7 @@ def challenge_host_list(request, challenge_host_team_pk):
 @throttle_classes([UserRateThrottle])
 @permission_classes((permissions.IsAuthenticated, HasVerifiedEmail))
 @authentication_classes((ExpiringTokenAuthentication,))
-def challenge_host_detail(request, challenge_host_team_pk, pk):
+def challenge_host_delete(request, challenge_host_team_pk, challenge_host_pk):
     try:
         challenge_host_team = ChallengeHostTeam.objects.get(
             pk=challenge_host_team_pk
@@ -162,7 +162,7 @@ def challenge_host_detail(request, challenge_host_team_pk, pk):
         return Response(response_data, status=status.HTTP_406_NOT_ACCEPTABLE)
 
     try:
-        challenge_host = ChallengeHost.objects.get(pk=pk)
+        challenge_host = ChallengeHost.objects.get(pk=challenge_host_pk)
     except ChallengeHost.DoesNotExist:
         response_data = {"error": "ChallengeHost does not exist"}
         return Response(response_data, status=status.HTTP_406_NOT_ACCEPTABLE)
@@ -202,8 +202,25 @@ def challenge_host_detail(request, challenge_host_team_pk, pk):
             )
 
     elif request.method == "DELETE":
-        challenge_host.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if challenge_host_team.created_by == request.user:
+
+            if (
+                challenge_host.user == request.user
+            ):  # when the user tries to remove himself
+                response_data = {
+                    "error": "You are not allowed to remove yourself since you are admin. Please delete the team if you want to do so!"
+                }  # noqa: ignore=E501
+                return Response(
+                    response_data, status=status.HTTP_406_NOT_ACCEPTABLE
+                )
+            else:
+                challenge_host.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            response_data = {
+                "error": "Sorry, you do not have permissions to remove this challenge host"
+            }
+            return Response(response_data, status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(["POST"])
@@ -254,45 +271,45 @@ def remove_self_from_challenge_host_team(request, challenge_host_team_pk):
         return Response(response_data, status=status.HTTP_401_UNAUTHORIZED)
 
 
-@api_view(["DELETE"])
-@throttle_classes([UserRateThrottle])
-@permission_classes((permissions.IsAuthenticated, HasVerifiedEmail))
-@authentication_classes((ExpiringTokenAuthentication,))
-def delete_challenge_host_from_team(request, challenge_host_team_pk, challenge_host_pk):
-    """
-    Deletes a challenge_host from a challenge_host_Team
-    """
-    try:
-        challenge_host_team = ChallengeHostTeam.objects.get(pk=challenge_host_team_pk)
-    except ChallengeHostTeam.DoesNotExist:
-        response_data = {"error": "ChallengeHostTeam does not exist"}
-        return Response(response_data, status=status.HTTP_406_NOT_ACCEPTABLE)
+# @api_view(["DELETE"])
+# @throttle_classes([UserRateThrottle])
+# @permission_classes((permissions.IsAuthenticated, HasVerifiedEmail))
+# @authentication_classes((ExpiringTokenAuthentication,))
+# def delete_challenge_host_from_team(request, challenge_host_team_pk, challenge_host_pk):
+#     """
+#     Deletes a challenge_host from a challenge_host_Team
+#     """
+#     try:
+#         challenge_host_team = ChallengeHostTeam.objects.get(pk=challenge_host_team_pk)
+#     except ChallengeHostTeam.DoesNotExist:
+#         response_data = {"error": "ChallengeHostTeam does not exist"}
+#         return Response(response_data, status=status.HTTP_406_NOT_ACCEPTABLE)
 
-    try:
-        challenge_host = ChallengeHost.objects.get(pk=challenge_host_pk)
-    except ChallengeHost.DoesNotExist:
-        response_data = {"error": "ChallengeHost does not exist"}
-        return Response(response_data, status=status.HTTP_406_NOT_ACCEPTABLE)
+#     try:
+#         challenge_host = ChallengeHost.objects.get(pk=challenge_host_pk)
+#     except ChallengeHost.DoesNotExist:
+#         response_data = {"error": "ChallengeHost does not exist"}
+#         return Response(response_data, status=status.HTTP_406_NOT_ACCEPTABLE)
 
-    if challenge_host_team.created_by == request.user:
+#     if challenge_host_team.created_by == request.user:
 
-        if (
-            challenge_host.user == request.user
-        ):  # when the user tries to remove himself
-            response_data = {
-                "error": "You are not allowed to remove yourself since you are admin. Please delete the team if you want to do so!"
-            }  # noqa: ignore=E501
-            return Response(
-                response_data, status=status.HTTP_406_NOT_ACCEPTABLE
-            )
-        else:
-            challenge_host.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-    else:
-        response_data = {
-            "error": "Sorry, you do not have permissions to remove this challenge host"
-        }
-        return Response(response_data, status=status.HTTP_401_UNAUTHORIZED)
+#         if (
+#             challenge_host.user == request.user
+#         ):  # when the user tries to remove himself
+#             response_data = {
+#                 "error": "You are not allowed to remove yourself since you are admin. Please delete the team if you want to do so!"
+#             }  # noqa: ignore=E501
+#             return Response(
+#                 response_data, status=status.HTTP_406_NOT_ACCEPTABLE
+#             )
+#         else:
+#             challenge_host.delete()
+#             return Response(status=status.HTTP_204_NO_CONTENT)
+#     else:
+#         response_data = {
+#             "error": "Sorry, you do not have permissions to remove this challenge host"
+#         }
+#         return Response(response_data, status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(["POST"])
